Renamed backend folders for clearer convention
