+++ driving-day-app-backend/app/apps.py
@@ -3,4 +3,4 @@ from django.apps import AppConfig
 
 class FsaeBackendAppConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
-    name = 'fsae_backend_app'
+    name = 'app'
