+++ driving-day-app-backend/app/ld_parser/file_utils.py
