+++ driving-day-app-backend/config/asgi.py
@@ -14,6 +14,6 @@ from django.core.asgi import get_asgi_application
 # import tracemalloc
 # tracemalloc.start()
 
-os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'fsae_backend.settings')
+os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'config.settings')
 
 application = get_asgi_application()
